Day 17

+++ src/day16/mod.rs
@@ -2,7 +2,7 @@ use std::collections::HashMap;
 
 use regex::Regex;
 
-pub fn solver(star: u8) -> fn(String) -> i32 {
+pub fn solver(star: u8) -> fn(String) -> i128 {
     match star {
         1 => star1,
         2 => star2,
@@ -10,17 +10,17 @@ pub fn solver(star: u8) -> fn(String) -> i32 {
     }
 }
 
-fn star1(input: String) -> i32 {
+fn star1(input: String) -> i128 {
     let re = Regex::new(r"Sue (\d+): (\w+): (\d+), (\w+): (\d+), (\w+): (\d+)").unwrap();
     let mut data = HashMap::new();
     let lines: Vec<_> = input.split("\n").collect();
     for line in lines.into_iter() {
         for cap in re.captures_iter(&line) {
             let mut sue = HashMap::new();
-            sue.insert(cap[2].to_string(), *&cap[3].parse::<i32>().unwrap());
-            sue.insert(cap[4].to_string(), *&cap[5].parse::<i32>().unwrap());
-            sue.insert(cap[6].to_string(), *&cap[7].parse::<i32>().unwrap());
-            data.insert(*&cap[1].parse::<i32>().unwrap(), sue);
+            sue.insert(cap[2].to_string(), *&cap[3].parse::<i128>().unwrap());
+            sue.insert(cap[4].to_string(), *&cap[5].parse::<i128>().unwrap());
+            sue.insert(cap[6].to_string(), *&cap[7].parse::<i128>().unwrap());
+            data.insert(*&cap[1].parse::<i128>().unwrap(), sue);
         }
     }
 
@@ -48,17 +48,17 @@ fn star1(input: String) -> i32 {
     panic!("Sue not found!");
 }
 
-fn star2(input: String) -> i32 {
+fn star2(input: String) -> i128 {
     let re = Regex::new(r"Sue (\d+): (\w+): (\d+), (\w+): (\d+), (\w+): (\d+)").unwrap();
     let mut data = HashMap::new();
     let lines: Vec<_> = input.split("\n").collect();
     for line in lines.into_iter() {
         for cap in re.captures_iter(&line) {
             let mut sue = HashMap::new();
-            sue.insert(cap[2].to_string(), *&cap[3].parse::<i32>().unwrap());
-            sue.insert(cap[4].to_string(), *&cap[5].parse::<i32>().unwrap());
-            sue.insert(cap[6].to_string(), *&cap[7].parse::<i32>().unwrap());
-            data.insert(*&cap[1].parse::<i32>().unwrap(), sue);
+            sue.insert(cap[2].to_string(), *&cap[3].parse::<i128>().unwrap());
+            sue.insert(cap[4].to_string(), *&cap[5].parse::<i128>().unwrap());
+            sue.insert(cap[6].to_string(), *&cap[7].parse::<i128>().unwrap());
+            data.insert(*&cap[1].parse::<i128>().unwrap(), sue);
         }
     }
